(PUP-7368) Add ability to declare function argument mismatch handlers
This commit adds a alternative `dispatch` to a function. It is named
`argument_mismatch`. The function that it calls must produce a String
that represents a message describing that the matched arguments are in
fact in error. An ArgumentError is then raised when the function returns
that string.

An `argument_mismatch` dispatcher will be excluded when the default
argument mismatch handler describes viable options. It is also excluded
from the list of dispatchers returned from the `Dispatcher`.

A function may have zero to many `argument_mismatch` declarations and
when none of them matches, the default mismatch describer still kicks
in.