(PUP-7273) Improve hiera function error message when path is missing
Before this commit, a missing path declaration in hiera.yaml resulted in
an error explaing that the options hash sent to a data_hash function that
required a path had an incorrect size (0 instead of 1). While correct,
that message was not very informative. This commit leverages the new
`argument_mismatch` dispatcher added by PUP-7368 to generate more useful
errors when the path is missing.