(PUP-6854) Add paths to skip for launchd
On macOS 10.12 and higher, the file `/System/Library/LaunchDaemons/com.apple.jetsamproperties.Mac.plist` exists which is not actually a service.  This would generate the following warning on every single Puppet run:

```
Warning: The /System/Library/LaunchDaemons/com.apple.jetsamproperties.Mac.plist plist does not contain a 'label' key; Puppet is skipping it
```

As this file (and perhaps others in the future) will never be a service, this adds a list of files which we will now skip from reading at all.