(PUP-7370) Ensure value returned from provider is checked only once.
A data provider function that declares a return type, must use a type
that is assignable to `Puppet::LookupValue` or, in case the function is
a `data_hash`, to `Hash[Puppet::LookupKey,Puppet::LookupValue]`. This
commit ensures that an error is raised if the return type is incorrect.

In this commit, it's also ensured that the lookup framework will rely
on function to do return type validation when all of its dispatchers
declares a correct return type.