(PUP-7371) Fix bug with missing undef in type mismatch description
The mismatch describer "forgot" that a type was optional when describing
it and in some cases simply dropped that type out of the description.
This commit ensures that the optional status is remembered and reported.

The commit also improves the type generalization of the `Variant` type
with respect to `Undef` and `Optional` so that a type that contains
either one becomes an `Optional` type that contains neither.

A side effect from fixing the above was that some tests needed to be
adjusted.