(PUP-7281) Add ability to declare options as an entry in Hiera defaults
This commit makes it possible to declare an `options` hash under the
hiera.yaml `defaults` entry. That hash will then be used by all entries
beneath `hierarchy` that does not declare an `options` entry of their
own. The `options` hash is never merged. If an entry declares it, it
replaces the default.