(PUP-7330) Ensure that compiler always initializes loaders and Pcore
Under some very rare circumstances, such as when the running the lookup
CLI command and the only available data provider for Hiera was a custom
backend written in Ruby, the Pcore would not be initialized by the
compiler since no loader was ever requested. This in turn, resulted in
that the `Puppet::LookupKey` and `Puppet::LookupValue` types never got
initialized.

This commit changes the compiler so that it performs the task of
initializing loaders and Pcore during its own initialization instead of
performing it lazily when the loaders are first requested.

The change in behavior broke a couple of tests that made assumptions on
the deferred loader initialization. This commit fixes those tests as
well.