(PUP-7237) Restore BEAKER_HOSTS env variable support
Commit fb9ab3b18 removed the ability to specify a pre-generated beaker
hosts config file using the `BEAKER_HOSTS` environment variable which is
used in ci-job-configs. Instead the rake task was combining the default
master platform and the `TEST_TARGET`, which already contains the master
platform resulting in a BHG string like:

    redhat7-64m-redhat7-64m-redhat6-64a

However, this causes the beaker run to fail because there are two master
platforms specified.

This commit restores the ability to specify a pre-generated beaker hosts
config file using `BEAKER_HOSTS`.

[skip ci]