(PUP-7237) Simplify host file resolution
Previously, the `gen_hosts` and `hosts_opt` methods checked if an
explicit host file was specified in the options hash, and if so, did a
noop. It also wasn't clear what the single source of truth was for
the host file.

When testing against `ci:test:gem` or `ci:test_against_preserved_hosts`,
we already know what host file to use. In that case, don't
depend on the gen_hosts task, and pass the host file in the options
hash to `beaker_test`, which always takes precedence over
environment settings.

In all other tasks that call `beaker_test`, depend on the `gen_hosts`
task to resolve the host file. The result of that task is to set
ENV['HOSTS'], which is the single source of truth for the host file,
unless overridden by the options hash.

This also removes the description from the `gen_hosts` task so that it
doesn't appear in the `rake -T` output, since it's not supposed to be
executed externally.

[skip ci]