(PUP-7334) Add Hiera default_hierarchy to module hiera.yaml
This commit adds the ability to declare a second hierarchy named
`default_hierarchy` to a hiera.yaml file that resides in a module root.
This hierarchy is only allowed in modules and will be consulted when, and
only when, a normal lookup does not find a value.

The `default_hierarchy` is insensitive to `lookup_options` declared in
the regular `hierarchy` and vice versa.

The `default_hierarchy` is also insensitive to any merge directives
passed in as a parameter in the lookup call.