(PUP-7372) Confine Hiera 3 to global and environment layer
This commit ensures that a data file that is reached using a Hiera 3
configuration and that in turn contains interpolation expressions that
use the 'hiera' or 'lookup' function, will have that lookup confined
to the global layer unless an environment data provider, version 5
is present. It also ensures that when an interpolation that originates
from Hiera 3 is calling out, that the special `Hiera::Scope` class
is unwrapped.

This was not the case earlier which resulted in that the special class
`Hiera::Scope` was passed on to the environment and module layers which
resulted in errors as they tried to access the `Scope#environment`
method which is not present in that class. Other side-effects of passing
that scope would be to make the special variables that it contains
available to data configured using Hiera 5.