(PUP-7328) Ensure that the 'settings' Class is created only once
Before this commit, each compilation would make an attempt to create
and add the 'settings' Class to the list of known resource types in the
environment. When more than one compilation was performed during the
life span of an environment, a warning "Class 'settings' is already
defined" would appear in the log. This is now changed so that the
compiler instead reuses the Class if it is already present in the
environment.