(PUP-6984) Set aliases when compiling
Puppet did not set explicit aliases when resources were added to the
catalog. This is a problem because it makes it impossible to provide
validation of references to resources using an alias.

This commit processes given aliases in a resource such that a single
alias, or a set of aliases in an array can be given.