(PUP-5659) Validate relationships formed via meta-parameter
Before this, a relationship formed via a meta-parameter was not
validated. This was bad because it meant that errors in manifests were
not detected during compilation and instead blew up much later when
applying the catalog.

Before this it was possible to use strict to control if a warning or
error should be emitted. Now a reference to a non existing resource is
always an error.