(PUP-7391) Ensure that the BaseLoader can handle self recursion
This commit ensures that the BaseLoader#internal_load method is prepared
for the case when its call to the parent loader initiates another load
of the same entity that it is currently loading. When this happens, the
resulting recursion loads the type and when it returns, the type will
be present in the loader. A check is therefore inserted to prevent an
attempt to add the entity once again, since that would result in an
error.