(PUP-6084) Make reference to Class[<upper case name>] be an error
This removes the deprecated (since Puppet 4.6.0) construct of
referencing a class with an upper case name, for example Class[Foo] by
making it always be an error.

This is removed since Foo evaluates to a type reference, and it may be a
type alias or a Resource. It may also be a type that requires
parameters. The use of an upper case name thus creates unssesary type
loading, and potentially a mysterious error.

The correct reference is always a String, and this string is always in
lower case.