(PUP-7402) Make Puppet references honour trailing slash in file titles
Before this, references to File resources with and without a trailing
slash would be considered different. This was caused by:

* avoidance to search the catalog based on a resource uniqueness key if
there were no alias in the catalog (the trailing slash does not add an
alias). This was strange, as it could work if some other resource used
an alias.
* Not including the default values for resource attributes derived from
the title when constructing aliases and references.

As a consequence several tests needed to be updated.
* Because of mocking and not expecting calls to stubs
* Assumption that a duplication would not be found based on title (i.e.
relied on the faulty behavior now corrected).

Method comments added or updated to help the next person trying to
figure out how references and aliases work (it is a mess).

Cleanup was made of commented out code from PUP-4947.