(PUP-5756) convert logs to UTF-8 and handle invalid encoding
This commit updates Puppet::Util::Log to convert new messages to UTF-8 with
Puppet::Util::CharacterEncoding. This will be useful for eventlog and other
destinations which do not perform conversion on their own.

Previously if a log event with a message containing invalid encoding was
generated, the handler would likely fail or behave unpredictably, up to
silently swalling the error and possibly never logging the event. Now, in
addition to trying to convert to UTF-8, if we detect that a message to be
logged contains invalid encoding we update the logged message to include a
warning about the invalid string, as well as the `dump` of the string, which is
the string with invalid bytes escaped for inspection. To help the receiver
identify the source of the issue, the last ten elements in the stack backtrace
are included in the augmented message.

Signed-off-by: Moses Mendoza <moses@puppet.com>