(PUP-7259) Handle deserializing report responses
Previously we could not deserialize json responses for report save
requests, because the rest terminus would attempt to call
`Array#from_data_hash`, which obviously doesn't exist.

This failure showed that the format API assumes the `intern` method will
short-circuit if `data` is already an instance of `klass`.

This commit adds tests for yaml, pson and json, and updates the json
format to behave like pson and yaml.