(PUP-7418) Fix problem with interpolated lookups
Before this commit, an interpolation using a module prefixed key and a
function that in turn made a full lookup (alias, lookup, or hiera),
would reset the `module_name` in the lookup invocation. This would
normally go undetected since the value, when found, would end the
lookup. During a deep_merge however, when the sub-lookup was made from
global or environment layer, the lookup's desire to find more data in
the module layer would not be satisfied correctly unless the sub-lookup
was targeting the same module as the original lookup.

This commit ensures that a new invocation is used for the sub-lookup,
leaving the original invocation untouched.