(PUP-7420) Minor performance improvement for lookup_options
Work on this ticket revealed an unnecessary call to the method
LookupAdapter#retrieve_lookup_options when a call was first made to
resolve a qualified key and then later to resolve an unqualified key.
The latter call would then create a sub invocation and add some
addtional explain output, just to find the value that had already been
cached for the environment.

This commit ensures that the environment value is cached properly even
when the request is for the lookup_values in a module.