(PUP-7259) Temporarily default reports to PSON
In order to test these changes in CI, we need to revert to PSON when
sending reports against puppetserver versions that don't yet know about
JSON as a serialization format.

For now, set the default report format to PSON, but mark the test as
pending, so it's clear what the ultimate desired behavior is.