(PUP-7429) Change default string format for Regexp
This commit changes the default string format (the '%s' format) so that
a formatted regexp results in a string that isn't delimited with
slashes.

The reason for the change is that the slashes are language specific, and
as such, should be produced using the '%p' format (which they already
are). The '%s' format should produce a language agnostic representation
that is suitable to pass to the Regexp constructor.