(PUP-7429) Ensure that regexp using flags are formatted with (? syntax
This commit ensures that regexps that contains flags are formatted using
the `Regexp#to_s` method which puts flags in a group construct. Regexps
without flags will instead use the `Regexp#source` method. The reason
for not always using the `#to_s` method is that it will output all flags
with a dash in front of the disabled ones, even when all flags are
disabled.