(maint) Make shared type collections available to rspec context
It has been an annoyance for some time that the named type collections
declared in the rspec shared context 'types_setup' isn't availabel on
the context level, and hence only can be used inside of an example. That
in turn means that no collection can be used to drive the set of examples
that should run. Everything has to take place inside the example which
in turn means that all 'let' etc. only initializes once for all the
types.

This commit ensures that all type collections in the 'types setup'
context are availabel on both the context level and in the examples.