(PUP-7427) Store rich data as normal parameter of type Hash
This commit changes the way resource parameters with rich data values
are stored in the catalog. Instead of using a special 'ext_parameters'
entry, the value is stored as a two element hash containing a special
key in the normal 'parameters' entry. The special key is
'__pcore_type__' and its value is the name of the data type of the
value that is associated with the second key 'value' in the same hash.

Any client that reads this catalog and is aware of the special rich
data key, will deserialize the value correctly. Clients that are
unaware of the rich data will just see a hash.

The value will be the string representation of the original value
unless the original is an Array, Hash, or a Pcore Object, in which
case Pcore serialization will serialize the value into an array.