(PUP-7402) Update test now that trailing slash alias works
Previously, the test added two resources to the catalog where one had a
trailing slash. The test relied on `defined` returning false, as it did
before commit ae066f9.

As of ae066f9, puppet will now detect that the resource with the
trailing slash has the same name as the one without. So this test
started failing (since we never tried to add a duplicate resource).

This commit updates the test to unconditionally add the duplicate. It
also removes the app_management cases since that property is deprecated
and no longer controls any behavior. It also updates the regexp since
the error message changed slightly:

    Error: Evaluation Error: Error while evaluating a Resource
    Statement, Cannot alias File[/tmp/create-dir.foyI4D/] ...

[skip ci]