(PUP-7259) Use existing conventions to exclude binary
Previously, the binary format declared itself to always be supported,
but the supported_formats method removed binary from the list. So it
wasn't really supported by default.

This commit removes the supported? method for the binary format. The
format is supported if the underlying model, e.g.
Puppet::FileServing::Content, implements the required render and intern
methods, which by convention are named to_binary and from_binary.

This commit also removes the overridden render_multiple and
intern_multiple methods, since the default implementations provide an
error message.

Also note there was a copy/paste error in the exception message for
render_multiple.