(maint) Remove gem update --system and pin rainbow
Travis fails with the combination of bundler 1.7.6 and rubygems 2.6,
as they are not compatible:

    NoMethodError: undefined method `spec' for nil:NilClass

This commit removes the `gem update --system` since we are no longer
installing a particular version of bundler. The thinking is that Travis
provides consistent versions of rubygems and bundler in all supported
ruby versions that we support.

This commit also pins rainbow, as newer versions are not compatible[1] with
ruby 2.4 and bundler 1.14.0, leading to errors like:

    Gem::Ext::BuildError: ERROR: Failed to build gem native extension.

[1] https://github.com/sickill/rainbow/issues/48