(PUP-7258) Remove json_pure dependency
Commit 30e0f1bcd was supposed to remove the json_pure dependency, but
due to a conflict when rebasing wasn't. As a result, puppet was using
json_pure instead of native json in some cases. It also exposed other
places where json_pure was still a dependency:

* Bundler running within puppet uses its Gemfile which includes the
  .gemspec
* Bundler running within a module and a source checkout of puppet uses
  puppet's .gemspec
* Packaging uses project_data.yaml when building puppet as a gem
* Git acceptance on Solaris (10 and 11) install json_pure to satisfy
  puppet's dependencies.

All traces of json_pure have been removed so that the native version
already present in ruby 1.9.3 and up is used.