(PUP-7258) Use ENV['SIZE'] to control number of operations per run
Previously, the serialization benchmark performed ITERATIONS^2
operations, since it was being used as both the number of runs to
perform, and the number of times to serialize the catalog per run.

Since ITERATIONS is intended to control the number of runs, use SIZE to
control how many times we serialize the catalog. The default value is
100, but can be specified as follows (or as an ENV variable):

    $ bx rake benchmark:serialization SIZE=10