(PUP-5974) Remove IO.binwrite monkeypatch
IO.binwrite has been available since ruby 1.9.3, and since we only
support 1.9.3 and up, there is no reason to carry this monkey patch
anymore.

This commit only removes the monkey patch, but not the tests, so that we
can execute the tests against ruby's version of IO.binwrite, and make
sure we don't get different behavior when the monkey patch is removed.