(PUP-7070) Remove stacktrace from HTTP error responses
Previously, our REST error responses contained a `stacktrace` element,
and in 4.x this was changed to a deprecation warning so that the server
did not leak sensitive information to the caller.

This commit removes the `stacktrace` element from the response, and
related schema and API docs. Note the `stacktrace` element was optional
according to our schema, so this should not be backwards incompatible.