(PUP-7436) Fix bug causing ephemeral undef not to override parent scope
This commit fixes a bug in the ephemeral local scope of a
`Puppet::Parser::Scope` that caused key declared with an undef value to
become invisible. Instead of returning such a value, the lookup
traversal continued to the parent scope. The value of found in the
parent scope was then returned.