(PUP-3552) Return HTTP 400 if content-type is missing or body is invalid
Previously, if the client didn't specify a Content-Type header or the
server failed to parse the body of the request, e.g. report, then the
server would generate an HTTP 500, which is incorrect, since the client
is to blame for sending the bad request. Instead, return HTTP 400 Bad
Request in those cases.