(PUP-3552) Return HTTP 415 for unsupported content types
Previously, puppet returned an HTTP 500 error if the client submitted
content whose type was:

  1) unknown to the server
  2) known to the server, but the model doesn't support it, e.g. reports
  don't support the `dot` or `s` formats.
  3) known to the server, but not suitable, e.g. msgpack gem isn't
  installed in puppetserver.

The last case is important, because previously, the Request#format
method would return nil, which would cause `convert_from_format` to
fail ungracefully.

This commit adds a Request#formatter method to return the network
formatter corresponding to the client-specified Content-Type. Note
unsafe yaml formats are filtered out. The knowledge for whether the
model supports the format is moved to the IndirectedRoutes class.

The server will respond with HTTP 415 Unsupported Media Type if the
server doesn't know about the format; it knows about the format, but the
model doesn't support it; or the model supports it, but the format is
not suitable. This behavior ensures the server is future compatible with
newer agents -- so if the agent asks for a new content-type, but the
server doesn't support it, then the agent can downgrade, even in cases
where the agents speaks first, e.g. when sending a report.

This commit exposed a bug in IndirectorTesting which implemented the
from_binary method, but not to_binary, so it wasn't truly suitable for
tests that exercised application/octet-stream. This commit fixes that.

We continue to return HTTP 406 if the client sends a list of "Accepts"
content types, but none of them are acceptable to the server.