(PUP-996) Close a relationship chain gap caused by an empty set
Before this change, a relationship expression of $x -> $y -> $z where $y
evaluates to an empty set resulted in $x -> [], and [] -> $z (which is
essentially a noop). This is bad because the expected result is that the
relationships instead are formed between $x and $z.

This commit changes this. If a user wants the earlier result, the
chained relationships should be written as separate relationships. As an
example $x -> $y; $y -> $z.