(maint) Remove duplicate definition of StaticLoader::BUILTIN_TYPES
For some reason (merge problem most likely) the constant BUILTIN_TYPES
declared in the StaticLoader was not used. Instead, a new array was
declared with the same content (Node had been added there, but not in
the original). This commit ensures that the Node type is added to the
BUILTIN_TYPES and that the copy is removed.