(PUP-6694) Replace hard-coded Pcore Data type with a type alias
The `PDataType` class, hardcoded to reflect all types that are `Data`
compatible, also introduces a parallel way of resolving types which
forces complex conditional code the `PDataType` and `PVariantType` and
several declarations of DATA constants. This design does not scale and
prevents adding new types such as `RichData`.

This commit removes the `PDataType` class completely and instead adds
a `Data` alias which is initialized very early in the `StaticLoader`.
All special concern for the `PDataType` is also removed.