(pup-7498) add a scrub method
Add our own rendition of ruby's String.scrub. We defer to the original if it's
defined, which is only post-ruby 1.9.3.

Ruby 1.9.3 does not have a scrub method. Nor does using String#encode(<encoding>,
<encoding>, :replace => '?'...) work:
* Given a UTF-8 string, when Encoding::UTF-8 is passed as the
  first/second arguments, no replacement characters are inserted (ie, the call is
  a no-op).
* When Encoding::BINARY is passed as second argument any non-ASCII characters
  are _always_ replaced - even when the UTF-8 string was valid!

So instead we effectively make our own scrub, and only call to it when we need it.

Signed-off-by: Moses Mendoza <moses@puppet.com>