(PUP-996) Make a relationship with empty collection close the gap
Before this, if a collection returned an empty set, and the collection
was in the middle of a chain of relationship expressions, the result
would be that the chain was broken.

Now, the gap created by an empty collection will work as if the
collector was not present. For example

Notify[a] -> Notify <| tag == 'toothfairy' |> -> Notify[b]

Would result in Notify[a] -> Notify[b].

This is implemented by using an adapter on a Collection. Collections
are lazy and the evaluator needs to delegate the decision to use the
left hand side result until the collection has been evaluated.
The adapter value is picked up in the Relationship class if the
collection is empty. It then finds the leftmost non empty result to use
instead of the gap.

In addition this required adding the notion of processing the left or
the right side of the relationship.