(PUP-7484) Move resource defaults initialization to initializer
In PUP-25, an attempt was made to add another phase to the compiler that
was responsible for setting default values. That was not sufficient since
new resources were created after the fact.

The PR for PUP-7484 that went in prior to this one attempted to fix this
by setting resource defaults on all resources created after the phase
where defaults had been set. This wasn't good enough either, since in
some cases, resources are created during evaluate_main phase.

This commit uses a simpler and more straight forward approach. Defaults
are now assigned in the Puppet::Parser::Resource#initialize method. The
method has an additional and optional argument that enables this behavior
to be turned off (needed when creating resources for the sake of
overriding parameter defaults).