(PUP-7464) Make json serialization handle :undef Symbol
Before this, an :undef symbol nested in an Array or Hash used as a
resource's parameter value would serialize the Ruby `:undef` symbol.

The support for rich data checks if the value is compliant with Json an
if not transforms the entire Array or Hash into a String.

This commit modifies the "to_pson" such that an `:undef` value is
transformed to `nil` (which can be represented in JSON). Prior to this
commit, only Array's where scanned for values in need of transformation,
now also Hashes are transformed.