(PUP-7492) Improve unit test bench for lookup interpolation
This commit changes the way interpolation tests are evaluated to more
closely resemble proper lookups. Here, a new LookupAdapter is used
that performs simple lookups in a provided hash.

This change was needed in order to write subsequent tests to assert
correct function when using nested interpolations with aliases and
lookups.