(PUP-7261) Add format charset option
Add a `charset` option so a format can specify its supported charset. By
default, a format doesn't have a charset, e.g. application/octet-stream, but a
charset can be specified on a per-format basis.

In theory, a format could support a list of charsets, and the network request
could negotiate the most suitable one based on the Accept-Charset header, like
we do for supported mime types. However, we don't need to support that
complexity, and we assume a format has at most one supported charset.