(PUP-7261) Append charset for textual content-types
Previously, the response forwarded the `type` parameter to the handler's
set_content_type method, where `type` could either be a String or Format. The
handler implementation for webrick/passenger/puppetserver, then had to call the
Format#format_to_mime method to resolve the mime type.

However, that approach won't work for charset, since it depends on the mime
type, the object being returned in the body of the response, and its encoding,
and the `format_to_mime` method doesn't have access to the body.

This commit flattens the content type, including charset if appropriate, in the
response before handing off to the handler. The `format_to_mime` method is
deprecated, since the handler should always receive the content-type, and
charset if appropriate.

If the request body is a string, then we transcode the body if its current
encoding doesn't match the desired charset.

This code assumes the format charset is always a valid Ruby encoding. This is
true for the most common charsets, e.g. UTF-8, CP1252, ISO-8859-1, etc.