(PUP-7261) Don't stub Zlib/Gzip in tests
Previously, the test stubbed Zlib inflate and Gzip reader methods, and made
assumptions that the decompressed/inflated strings were encoded in the ruby
default UTF-8.

Update the tests to use real Zlib/Gzip objects, which should not be a
problem for any recent ruby version. Adds 2 pending tests which show how
close is not always called on the inflater/reader objects. Makes a minor
change to the ZlibAdapter initializer so that the inflater can be passed
in. This won't affect any external consumers since the parameter is
optional.