(PUP-7110) Execute agent run in separate contract
Before this commit, it was not possible to restart/stop a puppet agent
service on Solaris from an agent run executed by that very service
without killing the agent run along with the service.
This was because services on Solaris run in so called contracts - a kind
of process groups - and during a service restart/stop all processes in
the contract are killed. Under normal circumstances the contract is
inherited during fork system call so the child process is a member of
the same contract as its parent process. This was the case of the agent
run executed by the puppet agent service on Solaris too, and that is
why the agent run could not restart/stop the puppet agent service
without killing itself.
So in this commit we simply make sure the agent run is executed in
a different contract than its parent (the puppet agent service) which
fixes the issue.