(PUP-5635) Change resolution of FQ name in scope to not become relative
Before this, a resolution of a name such as `::foo::bar` would delegate
the lookup of `bar` to the scope named `foo`. Thus, if `bar` was not
present in `foo`, it would lookup in the parent of `foo` (if it
inherits), and last search up the chain of scopes (typically node and
then top scope). This means a non existing FQ name would potentially
resolve to a node/top scope variable. (In the example here, if `$bar`
was set).

This is both baffling and wrong. The only relative resolutions that
should take place are references to unqualified variables (for example
`$bar`), and it should not take place for qualified variables such as
`$::bar` or `$::foo::bar`, or `$foo::bar`. The last since qualified
names (even if not fully qualified by being anchored by a leading `::`)
are not applied relative to namespaces. (Note that it is specified that
`foo::bar` === '::foo::bar`).