(PUP-5635) Add tests for expected scope FQ name lookup
Before this, one of the tests in `spec/integration/parser/scope_spec.rb`
specified the unwanted behavior (looking up `a::b` in nodescope if `a`
did not define `b`. That test is now modified to show the wanted
behavior (as specified by the PUP-5635 ticket).

Additional tests show that resolution of a FQ name does not take place
against node- or top-scope.