(maint) Remove worthless check for compiler passed in options
There is already a check that compiler is given as an argument.
It is stupid, but harmless to pass it as an option. There is no check
anywhere that scope options are valid. They are also not API.

Removing this random check as it does not really ensure better quality
code and costs a hash lookup and a conditional.