(maint) Remove dead code related to Scope#namespaces
Once upon a time, the `Array Scope#namespaces` was used to make a scope
exist in multiple namespaces. This was a seriously flawed idea akin to
multiple inheritance (without any safety measures) and it caused a lot
of confusion with what looked like non deterministic behavior (depended
on order of names in the namespace).

That was fixed, but the `Scope#namespaces` attribute and related logic
remained although the attribute is never read.

This removes this dead logic and fixes the tests where namespaces were
handled as if being a required option.

(The real namespace of a scope is obtained from the `source` assigned to
the scope. And there can be only one source).