(PUP-5635) Optimize use of Scope option :fq_leaf to be a parameter
This changes the implementation of the :fq_leaf option to instead be an
optional parameter in the `lookupvar` method.
This is better because the handling of the option requires an extra hash
merge for each qualified lookup and and two hash lookup for every
variable lookup. These are now instead a parameter pass, and local
variable references.