(PUP-5635) Use global hash for all qualified variables
With the change in behavior that all qualified names are only resolved
against the given name space (or inherited space for classes) it is
possible to optimize the lookup and keep a hash of fully qualified
names, and performing a lookup of the full name.

This commit adds this hash and modifies scope to make use of it.
As a consequence of this, it was possible to modify other methods
(inlining), and remove no longer used methods.

This shows a 5-10% performance increase on the fq variable benchmark.

One spec test was modified as the test assumed that a class named ""
represents the top scope (it is actually the scope set as top scope in
compiler) but the old resolution used the class named "" a.k.a "main".

This also enabled removing the extra parameter to control if a lookupvar
was a lookup of a leaf originating from a qualified name. The solution
in this commit is better because the former was a hack that only made
scope more complex.

Further improvements can now be made.