(PUP-7498) copy all struct values from Etc
Prior to this commit we errantly assumed all values from Etc were either
strings or arrays of strings. This was incorrect - Etc can return Fixnums as
well (and possibly others). Because we didn't handle this case, we dropped
these values when creating our new structs. This commit updates handling to
assign any existing non-String or Array values to their corresponding
fields in the new struct.

Signed-off-by: Moses Mendoza <moses@puppet.com>